Rename file section-column.vue to SectionColumn.vue
