+++ pages/about.vue
@@ -49,7 +49,7 @@
 import { documents, documentFetcher } from '~/services/document-fetcher'
 import HeroBanner from '@/components/slices/HeroBanner'
 import SectionSlice from '@/components/slices/Section'
-import SectionColumnSlice from '~/components/slices/section-column'
+import SectionColumnSlice from '@/components/slices/SectionColumn'
 
 export default {
   name: 'About',
